Added files via upload
These are files that I filled in skeleton code for. They don't work on their own without the rest of Professor Chang's code. They are meant to showcase my ability to work with scala and utilize functional programming techniques. They are NOT intended to give any advantage to future students and if you are a current student of this class then you should seek help from classmates and TA's and refrain from copying my work.